sphinx documentation for T1
